+++ src/components/AppNavigator.js
@@ -1,0 +1,8 @@
+import { createStackNavigator } from 'react-navigation-stack';
+import Home from '../components/Home/Home';
+
+const AppNavigator = createStackNavigator({
+  Home: { screen: Home }
+});
+
+export default AppNavigator;
